chore: move appMode to dapp-types and renaming files for consistency
