+++ libs/browser-utils/src/download/index.ts
@@ -1,4 +1,4 @@
-export * from './download-string';
-export * from './download-blob';
-export * from './string-into-blob';
-export * from './cached-fetch';
+export { default as cachedFetch } from './cachedFetch';
+export { default as downloadBlob } from './downloadBlob';
+export { default as downloadString } from './downloadString';
+export { default as stringIntoBlob } from './stringIntoBlob';
