--- libs/browser-utils/src/download/string-into-blob.ts
@@ -1,3 +1,0 @@
-export function stringIntoBlob(text: string, fileType = 'plan/text'): Blob {
-  return new Blob([text], { type: fileType });
-}
