+++ libs/browser-utils/src/download/stringIntoBlob.ts
@@ -1,0 +1,5 @@
+function stringIntoBlob(text: string, fileType = 'plan/text'): Blob {
+  return new Blob([text], { type: fileType });
+}
+
+export default stringIntoBlob;
