+++ libs/browser-utils/src/index.ts
@@ -1,6 +1,8 @@
 export * from './download';
 export * from './platform';
-export * from './retry-promise';
+export * from './retryPromise';
 export * from './storage';
 export * from './logger';
-export * from './promise-utils';
+export * from './executorWithTimeout';
+export { default as executorWithTimeout } from './executorWithTimeout';
+export { default as retryPromise } from './retryPromise';
