+++ libs/browser-utils/src/logger/index.ts
@@ -1,1 +1,2 @@
-export * from './logger-service';
+export * from './LoggerService';
+export { default as LoggerService } from './LoggerService';
