+++ libs/browser-utils/src/platform/index.ts
@@ -1,1 +1,2 @@
-export * from './get-platform-metadata';
+export * from './getPlatformMetaData';
+export { default as getPlatformMetaData } from './getPlatformMetaData';
