+++ libs/dapp-types/src/appMode.ts
