+++ libs/dapp-types/src/index.ts
@@ -5,11 +5,12 @@ export * from './ChainId';
 export * from './Currency';
 export * from './InteractiveFeedback';
 export * from './Props';
-export * from './utils';
-export * from './WalletId';
-export * from './WebbError';
 export * from './Storage';
 export { default as Storagge } from './Storage';
+export * from './WalletId';
+export * from './WebbError';
+export * from './appMode';
+export * from './utils';
 
 export const zeroAddress = '0x0000000000000000000000000000000000000000';
 export const ZERO = 'ZERO';
