+++ libs/fixtures-deployments/src/evm/vanchor.ts
@@ -1,7 +1,7 @@
 // Copyright 2022 @webb-tools/
 // SPDX-License-Identifier: Apache-2.0
 
-import { withLocalFixtures } from '@webb-tools/react-environment/app-mode';
+import { withLocalFixtures } from '@webb-tools/dapp-types/appMode';
 
 import { getCachedFixtureURI } from '..';
 import { cachedFetch } from '@webb-tools/browser-utils';
