+++ libs/react-environment/src/index.ts
@@ -1,6 +1,5 @@
 export * from './RouterProvider';
 export * from './store';
-export * from './app-mode';
 export * from './arkworks-proving-manager.worker';
 export * from './circom-proving-manager.worker';
 export * from './SettingProvider';
